Get plots from Figures
